requirements.txt
albumentations>=1.4.3
numpy>=1.26.4
opencv-python>=4.10.0
onnx>=1.15.0
onnxruntime>=1.18.0
Pillow>=10.3.0
pygame>=2.5.0
scikit-learn>=1.5.2
torch>=2.3.1
torchvision>=0.18.1
tqdm>=4.66.4
